backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
PyJWT==2.10.1
cryptography==41.0.7
supabase==1.0.3
anthropic==0.40.0
# sentence-transformers removed - using Gemini embeddings (no torch = faster deploys)
google-generativeai==0.3.1
PyPDF2==3.0.1
python-docx==1.1.0
numpy==1.24.3
apscheduler==3.10.4

# Stripe payment processing
stripe==7.4.0

# Gmail API dependencies
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Email parsing dependencies
beautifulsoup4==4.12.2
python-dateutil==2.8.2
aiohttp>=3.9.0

# Discovery Agent dependencies (browser-based grant scraping)
playwright>=1.40.0  # Also run: playwright install chromium
